refactor: move heavy functions out of utils
and make sure cli.py does not load them by default
